client: rename fd_logreadlist to fid_read_reqs, move to unifyfs.c
Renames fd_logreadlist to fid_read_reqs since the name "fd" was a
misnomer as it operates on file ids and not file descriptors. It
also moves the function from unifyfs-sysio.c to unifyfs.c to go
with fid_write which is effectively the counterpart write function.
