+++ client/src/unifyfs-sysio.c
@@ -43,12 +43,6 @@
 #include "unifyfs-internal.h"
 #include "unifyfs-sysio.h"
 #include "margo_client.h"
-#include "ucr_read_builder.h"
-#include "seg_tree.h"
-
-#ifndef MAX
-#define MAX(a, b) ((a) > (b) ? (a) : (b))
-#endif
 
 /* ---------------------------------------
  * POSIX wrappers: paths
@@ -564,7 +558,7 @@ ssize_t unifyfs_fd_read(int fd, off_t pos, void* buf, size_t count)
 
     /* execute read operation */
     ssize_t retcount;
-    int ret = unifyfs_fd_logreadlist(&req, 1);
+    int ret = unifyfs_fid_read_reqs(&req, 1);
     if (ret != UNIFYFS_SUCCESS) {
         /* failed to issue read operation */
         errno = EIO;
@@ -1179,7 +1173,7 @@ int UNIFYFS_WRAP(lio_listio)(int mode, struct aiocb* const aiocb_list[],
     }
 
     if (reqcnt) {
-        rc = unifyfs_fd_logreadlist(reqs, reqcnt);
+        rc = unifyfs_fid_read_reqs(reqs, reqcnt);
         if (rc != UNIFYFS_SUCCESS) {
             /* error reading data */
             ret = -1;
@@ -1210,671 +1204,6 @@ int UNIFYFS_WRAP(lio_listio)(int mode, struct aiocb* const aiocb_list[],
 }
 #endif
 
-/* order by file id then by file position */
-static int compare_index_entry(const void* a, const void* b)
-{
-    const unifyfs_index_t* ptr_a = a;
-    const unifyfs_index_t* ptr_b = b;
-
-    if (ptr_a->gfid != ptr_b->gfid) {
-        if (ptr_a->gfid < ptr_b->gfid) {
-            return -1;
-        } else {
-            return 1;
-        }
-    }
-
-    if (ptr_a->file_pos == ptr_b->file_pos) {
-        return 0;
-    } else if (ptr_a->file_pos < ptr_b->file_pos) {
-        return -1;
-    } else {
-        return 1;
-    }
-}
-
-/* order by file id then by offset */
-static int compare_read_req(const void* a, const void* b)
-{
-    const read_req_t* ptr_a = a;
-    const read_req_t* ptr_b = b;
-
-    if (ptr_a->gfid != ptr_b->gfid) {
-        if (ptr_a->gfid < ptr_b->gfid) {
-            return -1;
-        } else {
-            return 1;
-        }
-    }
-
-    if (ptr_a->offset == ptr_b->offset) {
-        return 0;
-    } else if (ptr_a->offset < ptr_b->offset) {
-        return -1;
-    } else {
-        return 1;
-    }
-}
-
-/* notify our delegator that the shared memory buffer
- * is now clear and ready to hold more read data */
-static void delegator_signal(void)
-{
-    LOGDBG("receive buffer now empty");
-
-    /* set shm flag to signal delegator we're done */
-    shm_data_header* hdr = (shm_data_header*)(shm_recv_ctx->addr);
-    hdr->state = SHMEM_REGION_EMPTY;
-
-    /* TODO: MEM_FLUSH */
-}
-
-/* wait for delegator to inform us that shared memory buffer
- * is filled with read data */
-static int delegator_wait(void)
-{
-    int rc = (int)UNIFYFS_SUCCESS;
-
-    /* specify time to sleep between checking flag in shared
-     * memory indicating server has produced */
-    struct timespec shm_wait_tm;
-    shm_wait_tm.tv_sec  = 0;
-    shm_wait_tm.tv_nsec = SHM_WAIT_INTERVAL;
-
-    /* get pointer to flag in shared memory */
-    shm_data_header* hdr = (shm_data_header*)(shm_recv_ctx->addr);
-
-    /* wait for server to set flag to non-zero */
-    int max_sleep = 5000000; // 5s
-    volatile int* vip = (volatile int*)&(hdr->state);
-    while (*vip == SHMEM_REGION_EMPTY) {
-        /* not there yet, sleep for a while */
-        nanosleep(&shm_wait_tm, NULL);
-        /* TODO: MEM_FETCH */
-        max_sleep--;
-        if (0 == max_sleep) {
-            LOGERR("timed out waiting for non-empty");
-            rc = (int)UNIFYFS_ERROR_SHMEM;
-            break;
-        }
-    }
-
-    return rc;
-}
-
-/* copy read data from shared memory buffer to user buffers from read
- * calls, sets done=1 on return when delegator informs us it has no
- * more data */
-static int process_read_data(read_req_t* read_reqs, int count, int* done)
-{
-    /* assume we'll succeed */
-    int rc = UNIFYFS_SUCCESS;
-
-    /* get pointer to start of shared memory buffer */
-    shm_data_header* shm_hdr = (shm_data_header*)(shm_recv_ctx->addr);
-    char* shmptr = ((char*)shm_hdr) + sizeof(shm_data_header);
-
-    /* get number of read replies in shared memory */
-    size_t num = shm_hdr->meta_cnt;
-
-    /* process each of our read replies */
-    size_t i;
-    for (i = 0; i < num; i++) {
-        /* get pointer to current read reply header */
-        shm_data_meta* rep = (shm_data_meta*)shmptr;
-        shmptr += sizeof(shm_data_meta);
-
-        /* get pointer to data */
-        char* rep_buf = shmptr;
-        shmptr += rep->length;
-
-        /* get start and end offset of reply */
-        size_t rep_start = rep->offset;
-        size_t rep_end   = rep->offset + rep->length;
-
-        /* iterate over each of our read requests */
-        size_t j;
-        for (j = 0; j < count; j++) {
-            /* get pointer to read request */
-            read_req_t* req = &read_reqs[j];
-
-            /* skip if this request if not the same file */
-            if (rep->gfid != req->gfid) {
-                /* request and reply are for different files */
-                continue;
-            }
-
-            /* same file, now get start and end offsets
-             * of this read request */
-            size_t req_start = req->offset;
-            size_t req_end   = req->offset + req->length;
-
-            /* test whether reply overlaps with request,
-             * overlap if:
-             *   start of reply comes before the end of request
-             * AND
-             *   end of reply comes after the start of request */
-            int overlap = (rep_start < req_end && rep_end > req_start);
-            if (!overlap) {
-                /* reply does not overlap with this request */
-                continue;
-            }
-
-            /* this reply overlaps with the request, check that
-             * we didn't get an error */
-            if (rep->errcode != UNIFYFS_SUCCESS) {
-                /* TODO: should we look for the reply with an errcode
-                 * with the lowest start offset? */
-
-                /* read reply has an error, mark the read request
-                 * as also having an error, then quit processing */
-                req->errcode = rep->errcode;
-                continue;
-            }
-
-            /* otherwise, we have an error-free, overlapping reply
-             * for this request, copy data into request buffer */
-
-            /* start of overlapping segment is the maximum of
-             * reply and request start offsets */
-            size_t start = rep_start;
-            if (req_start > start) {
-                start = req_start;
-            }
-
-            /* end of overlapping segment is the mimimum of
-             * reply and request end offsets */
-            size_t end = rep_end;
-            if (req_end < end) {
-                end = req_end;
-            }
-
-            /* compute length of overlapping segment */
-            size_t length = end - start;
-
-            /* get number of bytes from start of reply and request
-             * buffers to the start of the overlap region */
-            size_t rep_offset = start - rep_start;
-            size_t req_offset = start - req_start;
-
-            /* if we have a gap, fill with zeros */
-            size_t gap_start = req_start + req->nread;
-            if (start > gap_start) {
-                size_t gap_length = start - gap_start;
-                char* req_ptr = req->buf + req->nread;
-                memset(req_ptr, 0, gap_length);
-            }
-
-            /* copy data from reply buffer into request buffer */
-            char* req_ptr = req->buf + req_offset;
-            char* rep_ptr = rep_buf  + rep_offset;
-            memcpy(req_ptr, rep_ptr, length);
-
-            /* update max number of bytes we have written to in the
-             * request buffer */
-            size_t nread = end - req_start;
-            if (nread > req->nread) {
-                req->nread = nread;
-            }
-        }
-    }
-
-    /* set done flag if there is no more data */
-    if (shm_hdr->state == SHMEM_REGION_DATA_COMPLETE) {
-        *done = 1;
-    }
-
-    return rc;
-}
-
-/* This uses information in the extent map for a file on the client to
- * complete any read requests.  It only complets a request if it contains
- * all of the data.  Otherwise the request is copied to the list of
- * requests to be handled by the server. */
-static void service_local_reqs(
-    read_req_t* read_reqs,   /* list of input read requests */
-    int count,               /* number of input read requests */
-    read_req_t* local_reqs,  /* list to copy requests completed by client */
-    read_req_t* server_reqs, /* list to copy requests to be handled by server */
-    int* out_count)          /* number of items copied to server list */
-{
-    /* this will track the total number of requests we're passing
-     * on to the server */
-    int local_count  = 0;
-    int server_count = 0;
-
-    /* iterate over each input read request, satisfy it locally if we can
-     * otherwise copy request into output list that the server will handle
-     * for us */
-    int i;
-    for (i = 0; i < count; i++) {
-        /* get current read request */
-        read_req_t* req = &read_reqs[i];
-
-        /* skip any request that's already completed or errored out,
-         * we pass those requests on to server */
-        if (req->nread >= req->length || req->errcode != UNIFYFS_SUCCESS) {
-            /* copy current request into list of requests
-             * that we'll ask server for */
-            memcpy(&server_reqs[server_count], req, sizeof(read_req_t));
-            server_count++;
-            continue;
-        }
-
-        /* get gfid, start, and length of this request */
-        int gfid         = req->gfid;
-        size_t req_start = req->offset;
-        size_t req_end   = req->offset + req->length;
-
-        /* lookup local extents if we have them */
-        int fid = unifyfs_fid_from_gfid(gfid);
-
-        /* move to next request if we can't find the matching fid */
-        if (fid < 0) {
-            /* copy current request into list of requests
-             * that we'll ask server for */
-            memcpy(&server_reqs[server_count], req, sizeof(read_req_t));
-            server_count++;
-            continue;
-        }
-
-        /* get pointer to extents for this file */
-        unifyfs_filemeta_t* meta = unifyfs_get_meta_from_fid(fid);
-        struct seg_tree* extents = &meta->extents;
-
-        /* lock the extent tree for reading */
-        seg_tree_rdlock(extents);
-
-        /* identify whether we can satisfy this full request
-         * or not, assume we can */
-        int have_local = 1;
-
-        /* this will point to the offset of the next byte we
-         * need to account for */
-        size_t expected_start = req_start;
-
-        /* iterate over extents we have for this file,
-         * and check that there are no holes in coverage,
-         * we search for a starting extent using a range
-         * of just the very first byte that we need */
-        struct seg_tree_node* first;
-        first = seg_tree_find_nolock(extents, req_start, req_start);
-        struct seg_tree_node* next = first;
-        while (next != NULL && next->start < req_end) {
-            if (expected_start >= next->start) {
-                /* this extent has the next byte we expect,
-                 * bump up to the first byte past the end
-                 * of this extent */
-                expected_start = next->end + 1;
-            } else {
-                /* there is a gap between extents so we're missing
-                 * some bytes */
-                have_local = 0;
-                break;
-            }
-
-            /* get the next element in the tree */
-            next = seg_tree_iter(extents, next);
-        }
-
-        /* check that we account for the full request
-         * up until the last byte */
-        if (expected_start < req_end) {
-            /* missing some bytes at the end of the request */
-            have_local = 0;
-        }
-
-        /* if we can't fully satisfy the request, copy request to
-         * output array, so it can be passed on to server */
-        if (!have_local) {
-            /* copy current request into list of requests
-             * that we'll ask server for */
-            memcpy(&server_reqs[server_count], req, sizeof(read_req_t));
-            server_count++;
-
-            /* release lock before we go to next request */
-            seg_tree_unlock(extents);
-
-            continue;
-        }
-
-        /* otherwise we can copy the data locally, iterate
-         * over the extents and copy data into request buffer,
-         * again search for a starting extent using a range
-         * of just the very first byte that we need */
-        next = first;
-        while ((next != NULL) && (next->start < req_end)) {
-            /* get start and end of this extent (reply) */
-            size_t rep_start = next->start;
-            size_t rep_end   = next->end + 1;
-
-            /* get the offset into the log */
-            size_t rep_log_pos = next->ptr;
-
-            /* start of overlapping segment is the maximum of
-             * reply and request start offsets */
-            size_t start = rep_start;
-            if (req_start > start) {
-                start = req_start;
-            }
-
-            /* end of overlapping segment is the mimimum of
-             * reply and request end offsets */
-            size_t end = rep_end;
-            if (req_end < end) {
-                end = req_end;
-            }
-
-            /* compute length of overlapping segment */
-            size_t length = end - start;
-
-            /* get number of bytes from start of reply and request
-             * buffers to the start of the overlap region */
-            size_t rep_offset = start - rep_start;
-            size_t req_offset = start - req_start;
-
-            /* if we have a gap, fill with zeros */
-            size_t gap_start = req_start + req->nread;
-            if (start > gap_start) {
-                size_t gap_length = start - gap_start;
-                char* req_ptr = req->buf + req->nread;
-                memset(req_ptr, 0, gap_length);
-            }
-
-            /* copy data from local write log into request buffer */
-            char* req_ptr = req->buf + req_offset;
-            off_t log_offset = rep_log_pos + rep_offset;
-            size_t nread = 0;
-            int rc = unifyfs_logio_read(logio_ctx, log_offset, length,
-                                        req_ptr, &nread);
-            if (rc == UNIFYFS_SUCCESS) {
-                if (nread < length) {
-                    /* account for short read by updating end offset */
-                    end -= (length - nread);
-                }
-                /* update max number of bytes we have filled in the req buf */
-                size_t req_nread = end - req_start;
-                if (req_nread > req->nread) {
-                    req->nread = req_nread;
-                }
-            } else {
-                LOGERR("local log read failed for offset=%zu size=%zu",
-                       (size_t)log_offset, length);
-                req->errcode = EIO;
-            }
-
-            /* get the next element in the tree */
-            next = seg_tree_iter(extents, next);
-        }
-
-        /* copy request data to list we completed locally */
-        memcpy(&local_reqs[local_count], req, sizeof(read_req_t));
-        local_count++;
-
-        /* done reading the tree */
-        seg_tree_unlock(extents);
-    }
-
-    /* return to user the number of key/values we set */
-    *out_count = server_count;
-
-    return;
-}
-
-/*
- * get data for a list of read requests from the
- * delegator
- *
- * @param read_reqs: a list of read requests
- * @param count: number of read requests
- * @return error code
- * */
-int unifyfs_fd_logreadlist(read_req_t* in_reqs, int in_count)
-{
-    int i;
-    int read_rc;
-
-    /* assume we'll succeed */
-    int rc = UNIFYFS_SUCCESS;
-
-    /* assume we'll service all requests from the server */
-    int count = in_count;
-    read_req_t* read_reqs = in_reqs;
-
-    /* TODO: if the file is laminated so that we know the file size,
-     * we can adjust read requests to not read past the EOF */
-
-    /* if the option is enabled to service requests locally, try it,
-     * in this case we'll allocate a large array which we split into
-     * two, the first half will record requests we completed locally
-     * and the second half will store requests to be sent to the server */
-
-    /* this records the pointer to the temp request array if
-     * we allocate one, we should free this later if not NULL */
-    read_req_t* reqs = NULL;
-
-    /* this will point to the start of the array of requests we
-     * complete locally */
-    read_req_t* local_reqs = NULL;
-
-    /* attempt to complete requests locally if enabled */
-    if (unifyfs_local_extents) {
-        /* allocate space to make local and server copies of the requests,
-         * each list will be at most in_count long */
-        size_t reqs_size = 2 * in_count * sizeof(read_req_t);
-        reqs = (read_req_t*) malloc(reqs_size);
-        if (reqs == NULL) {
-            return ENOMEM;
-        }
-
-        /* define pointers to space where we can build our list
-         * of requests handled on the client and those left
-         * for the server */
-        local_reqs = &reqs[0];
-        read_reqs  = &reqs[in_count];
-
-        /* service reads from local extent info if we can, this copies
-         * completed requests from in_reqs into local_reqs, and it copies
-         * any requests that can't be completed locally into the read_reqs
-         * to be processed by the server */
-        service_local_reqs(in_reqs, in_count, local_reqs, read_reqs, &count);
-
-        /* bail early if we satisfied all requests locally */
-        if (count == 0) {
-            /* copy completed requests back into user's array */
-            memcpy(in_reqs, local_reqs, in_count * sizeof(read_req_t));
-
-            /* free the temporary array */
-            free(reqs);
-            return rc;
-        }
-    }
-
-    /* TODO: When the number of read requests exceed the
-     * request buffer, split list io into multiple bulk
-     * sends and transfer in bulks */
-
-    /* check that we have enough slots for all read requests */
-    if (count > UNIFYFS_MAX_READ_CNT) {
-        LOGERR("Too many requests to pass to server");
-        if (reqs != NULL) {
-            free(reqs);
-        }
-        return ENOSPC;
-    }
-
-    /* order read request by increasing file id, then increasing offset */
-    qsort(read_reqs, count, sizeof(read_req_t), compare_read_req);
-
-    /* prepare our shared memory buffer for delegator */
-    delegator_signal();
-
-    /* we select different rpcs depending on the number of
-     * read requests */
-    if (count > 1) {
-        /* got multiple read requests,
-         * build up a flat buffer to include them all */
-        flatcc_builder_t builder;
-        flatcc_builder_init(&builder);
-
-        /* create request vector */
-        unifyfs_Extent_vec_start(&builder);
-
-        /* fill in values for each request entry */
-        for (i = 0; i < count; i++) {
-            unifyfs_Extent_vec_push_create(&builder,
-                read_reqs[i].gfid, read_reqs[i].offset, read_reqs[i].length);
-        }
-
-        /* complete the array */
-        unifyfs_Extent_vec_ref_t extents = unifyfs_Extent_vec_end(&builder);
-        unifyfs_ReadRequest_create_as_root(&builder, extents);
-        //unifyfs_ReadRequest_end_as_root(&builder);
-
-        /* allocate our buffer to be sent */
-        size_t size = 0;
-        void* buffer = flatcc_builder_finalize_buffer(&builder, &size);
-        assert(buffer);
-
-        LOGDBG("mread: n_reqs:%d, flatcc buffer (%p) sz:%zu",
-               count, buffer, size);
-
-        /* invoke multi-read rpc */
-        read_rc = invoke_client_mread_rpc(count, size, buffer);
-
-        /* free flat buffer resources */
-        flatcc_builder_clear(&builder);
-        free(buffer);
-    } else {
-        /* got a single read request */
-        int gfid      = read_reqs[0].gfid;
-        size_t offset = read_reqs[0].offset;
-        size_t length = read_reqs[0].length;
-
-        LOGDBG("read: offset:%zu, len:%zu", offset, length);
-
-        /* invoke single read rpc */
-        read_rc = invoke_client_read_rpc(gfid, offset, length);
-    }
-
-    /* bail out with error if we failed to even start the read */
-    if (read_rc != UNIFYFS_SUCCESS) {
-        LOGERR("Failed to issue read RPC to server");
-        if (reqs != NULL) {
-            free(reqs);
-        }
-        return read_rc;
-    }
-
-    /*
-     * ToDo: Exception handling when some of the requests
-     * are missed
-     * */
-
-    /* spin waiting for read data to come back from the server,
-     * we process it in batches as it comes in, eventually the
-     * server will tell us it's sent us everything it can */
-    int done = 0;
-    while (!done) {
-        int tmp_rc = delegator_wait();
-        if (tmp_rc != UNIFYFS_SUCCESS) {
-            rc = UNIFYFS_FAILURE;
-            done = 1;
-        } else {
-            tmp_rc = process_read_data(read_reqs, count, &done);
-            if (tmp_rc != UNIFYFS_SUCCESS) {
-                rc = UNIFYFS_FAILURE;
-            }
-            delegator_signal();
-        }
-    }
-
-    /* got all of the data we'll get from the server,
-     * check for short reads and whether those short
-     * reads are from errors, holes, or the end of the file */
-    for (i = 0; i < count; i++) {
-        /* get pointer to next read request */
-        read_req_t* req = &read_reqs[i];
-
-        /* if we hit an error on our read, nothing else to do */
-        if (req->errcode != UNIFYFS_SUCCESS) {
-            continue;
-        }
-
-        /* if we read all of the bytes, we're done */
-        if (req->nread == req->length) {
-            continue;
-        }
-
-        /* otherwise, we have a short read, check whether there
-         * would be a hole after us, in which case we fill the
-         * request buffer with zeros */
-
-        /* get file size for this file */
-        off_t filesize_offt = unifyfs_gfid_filesize(req->gfid);
-        if (filesize_offt == (off_t)-1) {
-            /* failed to get file size */
-            req->errcode = ENOENT;
-            continue;
-        }
-        size_t filesize = (size_t)filesize_offt;
-
-        /* get offset of where hole starts */
-        size_t gap_start = req->offset + req->nread;
-
-        /* get last offset of the read request */
-        size_t req_end = req->offset + req->length;
-
-        /* if file size is larger than last offset we wrote to in
-         * read request, then there is a hole we can fill */
-        if (filesize > gap_start) {
-            /* assume we can fill the full request with zero */
-            size_t gap_length = req_end - gap_start;
-            if (req_end > filesize) {
-                /* request is trying to read past end of file,
-                 * so only fill zeros up to end of file */
-                gap_length = filesize - gap_start;
-            }
-
-            /* copy zeros into request buffer */
-            char* req_ptr = req->buf + req->nread;
-            memset(req_ptr, 0, gap_length);
-
-            /* update number of bytes read */
-            req->nread += gap_length;
-        }
-    }
-
-    /* if we attempted to service requests from our local extent map,
-     * then we need to copy the resulting read requests from the local
-     * and server arrays back into the user's original array */
-    if (unifyfs_local_extents) {
-        /* TODO: would be nice to copy these back into the same order
-         * in which we received them. */
-
-        /* copy locally completed requests back into user's array */
-        int local_count = in_count - count;
-        if (local_count > 0) {
-            memcpy(in_reqs, local_reqs, local_count * sizeof(read_req_t));
-        }
-
-        /* copy sever completed requests back into user's array */
-        if (count > 0) {
-            /* skip past any items we copied in from the local requests */
-            read_req_t* in_ptr = in_reqs + local_count;
-            memcpy(in_ptr, read_reqs, count * sizeof(read_req_t));
-        }
-
-        /* free storage we used for copies of requests */
-        if (reqs != NULL) {
-            free(reqs);
-            reqs = NULL;
-        }
-    }
-
-    return rc;
-}
-
 ssize_t UNIFYFS_WRAP(pread)(int fd, void* buf, size_t count, off_t offset)
 {
     /* equivalent to read(), except that it shall read from a given
@@ -1901,7 +1230,7 @@ ssize_t UNIFYFS_WRAP(pread)(int fd, void* buf, size_t count, off_t offset)
 
         /* execute read operation */
         ssize_t retcount;
-        int ret = unifyfs_fd_logreadlist(&req, 1);
+        int ret = unifyfs_fid_read_reqs(&req, 1);
         if (ret != UNIFYFS_SUCCESS) {
             /* error reading data */
             errno = EIO;
@@ -1983,7 +1312,6 @@ ssize_t UNIFYFS_WRAP(pwrite64)(int fd, const void* buf, size_t count,
 int UNIFYFS_WRAP(ftruncate)(int fd, off_t length)
 {
     /* check whether we should intercept this file descriptor */
-    int origfd = fd;
     if (unifyfs_intercept_fd(&fd)) {
         /* get the file id for this file descriptor */
         int fid = unifyfs_get_fid_from_fd(fd);
